mcdir: refactor dir impl, proto specifics
